Worky work work

+++ source/core/progress/sink.hpp
@@ -1,13 +1,13 @@
 #pragma once
 
-#include <cstdint>
+#include <cstddef>
 
 namespace progress {
 
 class Sink {
 public:
 
-	virtual void start(uint32_t resolution) = 0;
+	virtual void start(size_t resolution) = 0;
 	virtual void end() = 0;
 	virtual void tick() = 0;
 };
