+++ source/core/progress/std_out.hpp
@@ -7,14 +7,14 @@ namespace progress {
 class Std_out : public Sink {
 public:
 
-	virtual void start(uint32_t resolution);
+	virtual void start(size_t resolution);
 	virtual void end();
 	virtual void tick();
 
 private:
 
-	uint32_t resolution_;
-	uint32_t progress_;
+	size_t resolution_;
+	size_t progress_;
 
 	float step_;
 	float threshold_;
